Gui: move Win32-specific FileDialog code to own file
